Revert "Auth + delete recipe-lists"
